src/sentimentcomprehend/requirements.txt
boto3==1.21.38
Flask==2.1.1
splunk-opentelemetry[all]